Changes to wget parameters in run.sh to retry
In our Ansible automation we run a check before executing
the Hanlon container to determine if it is up and running before
launching the atftpd container.  This change will allow the removal
of that task and provide us the ability to use `docker-compose` in addition
to continuing to use Ansible docker automation.